fix: resolve all compile errors for CI build
- settings.gradle.kts: fix dependencyResolutionManagement
- build.yml: reorder steps, setup-gradle before wrapper
- C++: add missing includes (atomic, algorithm, cinttypes)
- C++: define VK_LAYER_EXPORT macro properly
- C++: fix VK_NULL_HANDLE passed to vkGetPhysicalDeviceMemoryProperties
- C++: fix format string mismatches (PRIu64)
- C++: fix signed/unsigned narrowing in JNI
- Kotlin: remove View.holder (not a SurfaceView)
- Kotlin: fix ShizukuRemoteProcess -> Shizuku.newProcess
- Kotlin: guard Tile.subtitle with API 29 check
- Kotlin: guard registerReceiver 3-arg with API 33 check
- Kotlin: add Build import to MainActivity
- Resources: add adaptive launcher icons (mipmap)

+++ app/src/main/cpp/framegen_jni.cpp
@@ -204,8 +204,8 @@ Java_com_framegen_app_engine_FrameGenEngine_nativeInit(
         return JNI_FALSE;
     }
 
-    uint32_t width = ANativeWindow_getWidth(g_engine.window);
-    uint32_t height = ANativeWindow_getHeight(g_engine.window);
+    uint32_t width = static_cast<uint32_t>(ANativeWindow_getWidth(g_engine.window));
+    uint32_t height = static_cast<uint32_t>(ANativeWindow_getHeight(g_engine.window));
     LOGI("Window: %ux%u", width, height);
 
     // Configure
